clean up after refactor
