+++ all_config.py
@@ -11,7 +11,7 @@ from trackers.creds import *
 trackers_to_update = ['Coal Mines']# ['Hydropower']# ['Gas Pipelines'] #['Iron & Steel']
 new_release_date = 'May_2025' # for within about page NEEDS TO BE FULL MONTH
 releaseiso = '2025-05'
-priority = ['gcmt'] # europe # NOTE NEEDS TO BE [''] to be skipped NEEDS TO BE mapname in map_tab internal
+priority = [''] # europe # NOTE NEEDS TO BE [''] to be skipped NEEDS TO BE mapname in map_tab internal
                     # africa
                     # integrated
                     # europe
