--- trackers/make_all_maps.py
@@ -1,29 +1,0 @@
-
-
-#########
-# STEP 1: Pull all data into a huge df (including about pages) (have local copies of ggit, ggitlng, goit, ggiteu, ggitlngeu)
-# STEP 2: Filter based on tracker, geo, and fuel (use what we use for tests)
-# STEP 3: Save all data downloads and send to April (for goget use two tabs, for egt pull in last egt tabs for GOGPT)
-# STEP 4: Do all adjustments on renamed data
-# STEP 5: Make map file by running script for all regions and fuels
-
-# do all data downloads africa, latam, asia, europe
-# technically ggit, and ggit lng got updated when egt was updated for egt 
-# so those about pages should reflect that egt release date not the global release date
-# is that also true for gogpt eu tho? i mean no... 
-########
-
-# print date
-# print new data file
-# print new release date
-# print all gspread keys being used and test each one
-# print all local files being used and test each one
-# to a scratchpad df save original length of files
-# also save original length after filtering by region
-# also save original length after filtering by rough country
-# also save original plants by region, and by rough country
-# also save capacity by rough country
-# also save capacity by region
-
-    
-
