+++ trackers/map_class.py
@@ -6,7 +6,7 @@ import json
 import subprocess
 import boto3
 from .helper_functions import save_to_s3, replace_old_date_about_page_reg, check_for_lists, rebuild_countriesjs, pci_eu_map_read, check_and_convert_float, remove_diacritics, check_rename_keys, fix_status_inferred, conversion_multiply, workaround_table_float_cap, workaround_table_units
-from .all_config import client_secret_full_path, gem_path, tracker_to_fullname, tracker_to_legendname, iso_today_date, gas_only_maps, final_cols, renaming_cols_dict, ggit_geojson, ggit_lng_geojson, new_release_date, gspread_creds, africa_countries, asia_countries, europe_countries, latam_countries, full_country_list
+from ..all_config import client_secret_full_path, gem_path, tracker_to_fullname, tracker_to_legendname, iso_today_date, gas_only_maps, final_cols, renaming_cols_dict, ggit_geojson, ggit_lng_geojson, new_release_date, gspread_creds, africa_countries, asia_countries, europe_countries, latam_countries, full_country_list
 import geopandas as gpd
 import numpy as np
 import gspread
@@ -752,9 +752,9 @@ class MapObject:
             else:
                 print(f'subnat not here for {tracker_obj.name}')
                 input('check which tracker is missing subnat')
-            print(f'Adding {tracker_sel} gdf to renamed_gdfs')
+            # print(f'Adding {tracker_sel} gdf to renamed_gdfs')
             renamed_gdfs.append(gdf)
-            input('Check it adds for gogpt eu')
+            # input('Check it adds for gogpt eu')
